refactor(interactions): migrate pagination to registry
Extracts the page:{...} button handler from interactionCreate.ts into
src/interactions/pagination.ts. Behavior and customId unchanged.
